Add GCStarter
I was presented with an issue where the Program Updater is used to call the program from the Start Menu, if the Updater detects the program is already running, it will silently terminate without calling the program.
To get around this, I created this batch script that will check to see if the program is running, if it is not, it will call the updater.

In the event the script detects the program is running it will not call the updater and will call the program itself. This ensures I can always open the program, and can easily open multiple instances of the program is I need to.

Please note: I have obscured the Executable name, and file paths.
If this script was applicable to you, [folder] [folder2] and [Program] can be inferred.